Add back `keys` property to regex
The returned regex object from `pathToRegex` used to include the modified
keys array as a `keys` property in v1.7.0. The current API is not entirely
ideal, since a keys array must be passed, we could instead use a more
functional API where no array is mutated.

In the future this should be changed that no array needs to be passed to
`pathToRegex`, adding it as a property is done for backwards compatibility.